update test for encounter
